nf BJS2-21463 int to long refactoring

+++ src/main/java/faang/school/godbless/calculator/Calculator.java
@@ -4,23 +4,23 @@ import java.util.List;
 
 public class Calculator {
 
-    public static int calculate(List<Integer> nums, Calculators calculator) throws IllegalArgumentException {
+    public static long calculate(List<Integer> nums, Calculators calculator) throws IllegalArgumentException {
         if (nums.isEmpty()) {
             throw new IllegalArgumentException("List is empty");
         }
 
-        int result = nums.get(0);
+        long result = nums.get(0);
         for (int i = 1; i < nums.size(); i++) {
             result = calculator.calculate(result, nums.get(i));
         }
         return result;
     }
 
-    public static int sum(List<Integer> nums) {
+    public static long sum(List<Integer> nums) {
         return calculate(nums, (a, b) -> a + b);
     }
 
-    public static int product(List<Integer> nums) {
+    public static long product(List<Integer> nums) {
         return calculate(nums, (a, b) -> a * b);
     }
 }
